+++ lua/plugins/lspsaga.lua
@@ -9,7 +9,4 @@ return {
     "nvim-treesitter/nvim-treesitter", -- optional
     "nvim-tree/nvim-web-devicons", -- optional
   },
-  config = function()
-    vim.keymap.set("n", "K", "<cmd>Lspsaga hover_doc<CR>", { noremap = true, silent = true })
-  end,
 }